refactor: 에러 처리 수정
에러처리를 미들웨어를 통해 하고
에러는 유틸에서 생성 관리하는 것으로 변경
